Track number of music files and size of collection to copy over; add dry run mode.

+++ sync.rb
@@ -41,6 +41,9 @@ OptionParser.new do |parser|
   parser.on('-p', '--preserve', "Preserve existing files (don't delete any songs already on the card).") do |preserve_files_arg|
     options[:preserve_files] = preserve_files_arg
   end
+  parser.on('-d', '--dry-run', "Calculate the number of files and amount of data to copy, but do not actually copy anything.") do |dry_run_arg|
+    options[:dry_run] = dry_run_arg
+  end
   parser.on('-h', '--help', 'Display this help.') do
     puts parser
     exit
@@ -58,39 +61,41 @@ sync_hotlist = !!options[:hotlist]
 
 raise 'Nothing to sync: albums, continuity, favourites, hotlist or essentials must be specified.' unless sync_favourites || sync_essentials || sync_albums || sync_hotlist || sync_continuity
 
-TARGET_PATH = options[:target] || raise('Target path must be specified.')
+unless options[:dry_run]
+  TARGET_PATH = options[:target] || raise('Target path must be specified.')
+  Dir.chdir TARGET_PATH
+end
 LIBRARY_PATH = options[:library] || raise('Library path must be specified')
 library = Library.get_latest(LIBRARY_PATH)
-
-Dir.chdir TARGET_PATH
 
 def create_initial_directories(library)
   artists = library.artists
   artist_names = artists.map(&:name)
   initials = artist_names.map { |a| a[0].upcase }.uniq.sort
-  # Delete anything that doesn't correspond to a valid initial.
-  Dir['*'].each do |item|
-    FileUtils.rm_r(item, force: true) unless @preserve_files || initials.include?(item)
-  end
-
-  initials.each do |initial|
-    initial_dir = TARGET_PATH + initial
-    if File.exist? initial_dir
-      raise "#{initial_dir} exists but is not a directory." unless File.directory? initial_dir
-    else
-      Dir.mkdir initial_dir
+  unless options[:dry_run]
+    # Delete anything that doesn't correspond to a valid initial.
+    Dir['*'].each do |item|
+      FileUtils.rm_r(item, force: true) unless @preserve_files || initials.include?(item)
     end
-    # Delete any artists that have been removed from the artist list.
-    artists_with_this_initial = artist_names.select { |a| a[0].upcase == initial }
-    Dir.chdir initial_dir
-    existing_artist_dirs = Dir['*']
-    existing_artist_dirs.each do |d|
-      FileUtils.rm_r(d, force: true) unless @preserve_files || artists_with_this_initial.include?(d.chomp '/')
+    initials.each do |initial|
+      initial_dir = TARGET_PATH + initial
+      if File.exist? initial_dir
+        raise "#{initial_dir} exists but is not a directory." unless File.directory? initial_dir
+      else
+        Dir.mkdir initial_dir
+      end
+      # Delete any artists that have been removed from the artist list.
+      artists_with_this_initial = artist_names.select { |a| a[0].upcase == initial }
+      Dir.chdir initial_dir
+      existing_artist_dirs = Dir['*']
+      existing_artist_dirs.each do |d|
+        FileUtils.rm_r(d, force: true) unless @preserve_files || artists_with_this_initial.include?(d.chomp '/')
+      end
     end
   end
 end
 
-create_initial_directories library
+create_initial_directories library unless options[:dry_run]
 
 if options[:artist]
   artists_to_sync = [library.find_artist_by_name(options[:artist])]
@@ -109,26 +114,31 @@ else
 
 end
 
+collection_size = 0
+file_count = 0
+
 artists_to_sync.each do |artist|
   puts "\nSyncing #{artist.name}"
 
-  artist_path = "#{TARGET_PATH}#{artist.name[0].upcase}/#{artist.name}"
-  Dir.mkdir artist_path unless Dir.exist?(artist_path)
-  Dir.chdir artist_path
+  unless options[:dry_run]
+    artist_path = "#{TARGET_PATH}#{artist.name[0].upcase}/#{artist.name}"
+    Dir.mkdir artist_path unless Dir.exist?(artist_path)
+    Dir.chdir artist_path
 
-  # Delete anything that's not longer needed.
-  existing_items = Dir['*']
-  existing_items.each do |item|
-    FileUtils.rm_r(item, force: true) unless @preserve_files || artist.has_album_named?(item)
-  end
+    # Delete anything that's not longer needed.
+    existing_items = Dir['*']
+    existing_items.each do |item|
+      FileUtils.rm_r(item, force: true) unless @preserve_files || artist.has_album_named?(item)
+    end
 
-  # Create essentials playlist if it's wanted.
-  if sync_essentials && artist.has_essentials_tracks?
-    playlist_filename = "The Essential #{artist.name}.m3u"
-    FileUtils.rm(playlist_filename) if Dir.exist?(playlist_filename) && !@preserve_files
-    File.open(playlist_filename, 'a') do |file|
-      artist.essentials_tracks.each do |track|
-        file.puts track
+    # Create essentials playlist if it's wanted.
+    if sync_essentials && artist.has_essentials_tracks?
+      playlist_filename = "The Essential #{artist.name}.m3u"
+      FileUtils.rm(playlist_filename) if Dir.exist?(playlist_filename) && !@preserve_files
+      File.open(playlist_filename, 'a') do |file|
+        artist.essentials_tracks.each do |track|
+          file.puts track
+        end
       end
     end
   end
@@ -144,13 +154,15 @@ artists_to_sync.each do |artist|
         (sync_hotlist && track.hotlist)
     end
     source_directory = "#{SOURCE_PATH}#{artist.name[0].upcase}/#{artist.name}/#{album.title}/"
-    target_directory = "#{TARGET_PATH}#{artist.name[0].upcase}/#{artist.name}/#{album.title}"
-    Dir.mkdir target_directory unless Dir.exist? target_directory
-    Dir.chdir target_directory
-    # Remove any existing files we don't want.
-    existing_files = Dir['*']
-    existing_files.each do |file|
-      FileUtils.rm(file) unless @preserve_files || files_to_sync.include?(file)
+    unless options[:dry_run]
+      target_directory = "#{TARGET_PATH}#{artist.name[0].upcase}/#{artist.name}/#{album.title}"
+      Dir.mkdir target_directory unless Dir.exist? target_directory
+      Dir.chdir target_directory
+      # Remove any existing files we don't want.
+      existing_files = Dir['*']
+      existing_files.each do |file|
+        FileUtils.rm(file) unless @preserve_files || files_to_sync.include?(file)
+      end
     end
     # Copy over the files we want.
     files_to_sync.each do |file_to_sync|
@@ -159,50 +171,63 @@ artists_to_sync.each do |artist|
         puts "Cannot sync #{track_source} because it does not exist."
         next
       end
-      if File.exist?(file_to_sync)
-        if FileUtils.identical?(track_source, file_to_sync)
-          puts "     #{file_to_sync} -- Source file is identical to target file: skipping copy."
-          next
+      file_count += 1
+      collection_size += File.size track_source
+      unless options[:dry_run]
+        if File.exist?(file_to_sync)
+          if FileUtils.identical?(track_source, file_to_sync)
+            puts "     #{file_to_sync} -- Source file is identical to target file: skipping copy."
+            next
+          else
+            puts "     #{file_to_sync} -- Source file is different from target file."
+          end
         else
-          puts "     #{file_to_sync} -- Source file is different from target file."
+          puts "     #{file_to_sync} doesn't exist at the target location."
         end
-      else
-        puts "     #{file_to_sync} doesn't exist at the target location."
+        puts "**** Copying #{track_source} to #{Dir.pwd}"
+        FileUtils.cp(track_source, '.')
       end
-      puts "**** Copying #{track_source} to #{Dir.pwd}"
-      FileUtils.cp(track_source, '.')
     end
-    if Dir.empty? target_directory
-      # Remove the album directory if there's nothing in it.
-      FileUtils.rm_r(target_directory, force: true)
-    else
-      # Copy any jpg image files over.
-      Dir["#{source_directory}/*.jpg"].each do |image|
-        FileUtils.cp image, target_directory
+    unless options[:dry_run]
+      if Dir.empty? target_directory
+        # Remove the album directory if there's nothing in it.
+        FileUtils.rm_r(target_directory, force: true)
+      else
+        # Copy any jpg image files over.
+        Dir["#{source_directory}/*.jpg"].each do |image|
+          collection_size += File.size image
+          FileUtils.cp image, target_directory
+        end
       end
     end
   end
-  FileUtils.rm_r(artist_path) if Dir.empty? artist_path
+  FileUtils.rm_r(artist_path) if !options[:dry_run] && Dir.empty?(artist_path)
 end
 
 # Create favourites playlist if it's wanted.
-favourites_playlist = "#{TARGET_PATH}Radio Russ.m3u"
-if sync_favourites
-  FileUtils.rm(favourites_playlist) if Dir.exist? favourites_playlist
-  File.open(favourites_playlist, 'a') do |file|
-    library.favourite_tracks.each do |track|
-      file.puts "#{track.artist.initial_dir}#{track.artist.name}/#{track.album.title}/#{track.filename}"
+unless options[:dry_run]
+  favourites_playlist = "#{TARGET_PATH}Radio Russ.m3u"
+  if sync_favourites
+    FileUtils.rm(favourites_playlist) if Dir.exist? favourites_playlist
+    File.open(favourites_playlist, 'a') do |file|
+      library.favourite_tracks.each do |track|
+        file.puts "#{track.artist.initial_dir}#{track.artist.name}/#{track.album.title}/#{track.filename}"
+      end
+    end
+  end
+
+  # Create hotlist playlist if it's wanted.
+  hotlist_playlist = "#{TARGET_PATH}Hotlist.m3u"
+  if sync_hotlist
+    FileUtils.rm(hotlist_playlist) if Dir.exist? hotlist_playlist
+    File.open(hotlist_playlist, 'a') do |file|
+      library.hotlist_tracks.each do |track|
+        file.puts "#{track.artist.initial_dir}#{track.artist.name}/#{track.album.title}/#{track.filename}"
+      end
     end
   end
 end
 
-# Create hotlist playlist if it's wanted.
-hotlist_playlist = "#{TARGET_PATH}Hotlist.m3u"
-if sync_hotlist
-  FileUtils.rm(hotlist_playlist) if Dir.exist? hotlist_playlist
-  File.open(hotlist_playlist, 'a') do |file|
-    library.hotlist_tracks.each do |track|
-      file.puts "#{track.artist.initial_dir}#{track.artist.name}/#{track.album.title}/#{track.filename}"
-    end
-  end
-end
+puts "\n"
+puts '======================================================='
+puts "#{file_count} music files; #{collection_size / 1024 / 1024 } MB (including art)."
